We scrollin' now
Press a key to scroll down a line. It works, but it's crude as heck and doesn't know when to stop. It's a start though.

Add TWW's keyboard code as a suitable placeholder for now. I intend to write my own.

Saved bytes by removing the alignment on the logo sprites, we now copy them to aligned memory to use them.