IRQ improvements
The interrupt is now stable* when the custom handler is called, doing away with the manual adjustment

* Except badlines / sprite-lines etc. To be solved later

Much simpler 'wait for frame' by just checking the raster line MSB rather than using a variable (also reduces interrupt length)

The CPU port is backed up and restored during interrupt so that the memory map used in the main thread won't cause the interrupt to crash.

The menu database script was modified heavily to output an ACME file instead of a binary. This simplified menu drawing signigicantly which will make it easier to implement the keyboard/joypad navigation.

Small changes to the string printing, but for some reason the colour is not being applied.

BSOD64: Set lower-case and disable SHIFT+C= (to change case)