Add initial colour data to lines
This adds the first pieces of packing colour data into the lines.

Instead of managing semantics on the C64 (such as with embedded format codes), the colour will be arranged by the lua scripts and the C64 just streams out colour RAM data.

Most lines will not need colour data (using the default text colour) so I specifically avoid using an extra byte to indicate the presence of colour data or not; the high-bit of the line-length is used for this purpose.

For lines with colour data, the first byte is special, allowing us to make a whole line as a particular colour (one of seven style classes). This is what's currently implemented and demoed in this commit.

The text-printing was simplified quite a bit with better loop flow.