Use custom character mapping
This packs the font into ASCII order (and adds the ASCII charcaters missing in the C64 ROM) but ultimately gives us the ability to use any custom characters we want, e.g. smart-quotes, and custom decorators (such as horizontal-bars).

This will probably also help with compression later down the line as we can freely re-organise the screen codes and the lua scripts will build the correct screen output.