Move off-screen rows into high-RAM
This is to save program code space. Ideally, I want to fit the outfit under $1000 (where SIDs must go).

This might require using some RAM just before $C000 for code that won't fit. I'm not sure how to structure this yet.

Currently, the outfit starts at $0800 due to exomizer depacking. The fast-loader should allow us to move this down to $0400. (VIC bank 3 is used for the screen)

The article data has been moved to $4000 (from $2000). This gives us 8K for a SID ($1000-$4000).