Rework memory map / fast-loading
This splits the code into shared (<$1000) and main code ($B000..$C000) since the SID area is unmovable (90% of SIDs load to $1000).

A load of constants were made local to minimize the amount of leakage in the symbol file needed to separate shared and main code.

The bootstrap and intro do one-way JSRs so as to allow their own code to be overwritten by what's loaded in.

We can't use the Exomized binary ATM because the self-extracing code overwrites our shared code. A different methodology will be needed.