with much kludging, fast-load+decrunching
This switches to pucrunch for the cruncher/fast-load-decruncher as I could not get Exomizer working.

Pucrunch does not allow combining of multiple files, so we only compress the main outfit code for now.

I may switch to Krill's loader as not being able to fast-load in the background is of no use to us long-term, but it has an awfully heavy initial payload (6-7KB) which will mandate a fast-loader in the bootstrap! -- help very much wanted.