Improvements to ASCII to C64 text conversion
We now better account for the variability of output line length vs. ASCII length, and test line-length based on the line+word combined, rather than separate.

A subtle difference is that the C64 string is not built piece by piece, but on-demand, accounting for the whole line contents rather than each piece added.

All of this is necessary to move toward managing word-to-word colour data, which we don't handle yet.

As made possible by the above changes, I also added an "space-I-space" contraction. Possibly a waste of two screen codes, but does look very nice and occurs enough to warrant it.