Colour spans (WIP)
This covers about 90% of the ground work to process and display colour-spans, i.e. colouring indvidual characters or words in a line; only full-line colour is currently supported.

The Line class in the Lua scripts was split out to its own file.

Heavy re-working of the read-screen line-printing routine. Many small optimisations, and the colour-data decoding was split out and expanded.