Solve issues around ASCII to C64 conversion
This solves some difficulties with converting a source ASCII line to both C64 screen codes and colour spans;

Before the colour data was built as each word was appended, but the colour data could not be rebuilt from the final line string;

We now embed escape codes (0x1b) into the ASCII source string to indicate changes in style-class when converting the line to C64 screen codes and colour spans.

Some architectural clean-up still needs to occur to make the best use of this change.