Implemented nested styles, now allowing for correct display of noun within bold within title styles for example.
A great deal of simplification was done, doing away with the Line class, instead just storing the ASCII lines (with embedded control codes) as per-line attributes can be encoded that way instead of through object properties.

There remains a bug with the colouring of the first title of an article (something to do with the loss of `line.default`?)

Added a "soft" style class via "((...))" markup. Triple brackets can be used to output regular brackets in the soft style.