Fix single colour lines
Previous changes (the removal of the `line.default` property) caused single colour lines to appear as 2 colours due to the trailing spaces counting towards the number of colours.

We now bleed the the last character's style class to the right, across all the trailing spaces.

This commit also moves style minimisation up a level to the article, rather than string level, as we want our encoding and compression methods to be applied as late as possible.