Compression:
Adds a `Compress` object that helps separate compression concerns from the straight conversion from ASCII to screen-codes.

A test byte-pair compression is done, but the result is not written to disk yet. Results are okay for a first pass, achieving almost 50% compression.

A _lot_ of engineering work is still needed to output the compressed binary (dictionary must be included), including all the work on the 6502 side to decode it.