tokenise / compress
This is a reworking of the compression method to tokenise all screen-codes. Instead of alloting tokens from $80 upwards, we now count the number of unique screen-codes in the article and assign the first tokens to cover these, so that if there are 56 unqiue screen codes in the article, all lines are recoded so that only bytes 1-56 are used. We can then use the maximum number of free tokens available for token-pair encoding (rather than being limited to 128 in all cases).

This commit only implements a test algorithm and doesn't change the decoder on the C64. There's likely a number of major logic flaws in the current algorithm that need to be ironed out.

This commit also adds ACME source output. It's easier and more flexible to write out an assembler file than a binary, and will allow us to shift some of the constant values to the assembler side, rather than duplicating these values in the assembler / Lua scripts.