Increase token-space to $FF
Before, it was being used as an escape, but since we enumarate all currently-used screen-codes in the original article, we don't need it.

We also strip the spaces from the article, and convert screen-codes to lower-case. This gives stronger compression, but this will need a lot more implementation logic before it can be used.