Bit-pack spaces!
If you remove the spaces from a sentence, there will be less entropy with character-pairs (=better compression). The hard part is putting the spaces back in afterwards.

How do you add spaces back in though, when "the new" might tokenise the "e" & "n" together (without the space)?

Answer: We bit-pack the spaces into bytes and consume a byte of spaces as we go. As the tokens unpack, we will eventually get a literal. Before printing the literal, we check out a byte of spaces -- 0 = space, 1 = solid -- and print all contiguous spaces until we reach a solid.

Whenever 8 chars have been processed, we read another byte of spaces from the input -- the tricky part is encoding the stream so that the packed spaces are exactly in the right place where they will be *assumed* to be; no bits are ever wasted on "and now, some spaces".

The algorithm on the PC-side to put the bit-packed spaces in the right places was hard work to conceptualise but it boils down to "for a token that expands to n literals, we will need as many bit-packed space bytes as contains n *non-spaces*.

This whole thing is so crazy, I don't believe anybody else has done it before, at least on the C64.