Remove the address overhang used by the C64 article screen.
At an earlier point the top of the article was detected by the address of the index in the list of line-lengths under-flowing, but it's done by the vertical scroll position in pixels now.

The 'header' of the article needs major changes to support footnotes, so this is a prepatory change to support having more 'up-front' meta-data rather than relying on fixed sized tables.