Minor optimisations to C64 line-printing
Removed the blanking of a line before the new line is printed. The 'blanking' of the line is now done according to the line's contents; if a line is blank (length = 0), it is coloured invisible rather than printed as spaces (faster).

If a line is text, the decompression routine pre-fills the buffer with spaces. This prevents screen-garbage but may require further optimisation. If we were to fill the buffer backwards, we could fill spaces toward zero.