Convet line storage to tables...
It looks like for access to footnote metadata we'll have to merge the compressor back into Article rather than over-complicate the cross-communication.

This commit changes the Article's line-storage from just strings to tables with a "source" element; this is to make them compatible with the way the compressor stores lines so that the two can be merged.