Do not scroll row 25
This commit does nothing more than introduce another interrupt that resets the scroll regsiter on row 25 (where a fixed status bar will be). No stabilisation is done, yet, so it freaks out when you scroll.

Scrolling still uses row 25, so the bottom line changes as you scroll. Further updates will change this. Raster stabilisation is difficult, so I'm doing this in small steps.

We could use sprites in the border to draw our status bar, but our current approach is much more useful when we get to displaying footnotes.