Optimisations to scrolling
This moves the before / after screen buffers into the zero-page so that the screen scroll saves cycles in the copy loops.

Even though a zero-page buffer was being partially used previously, an oversight in register use meant that the wrong opcodes were being used and no time was being saved.

With the addtion of the two zero-page buffers and the opcode fixes to the copy loop, about 480 cycles are saved when scrolling up.

This timesave amounts to ~7.6 scanlines; not a lot, but will go a long way when SID music is added, which apparently can be as much as 20 lines!