Use environment variables for addresses?
Not for now, but we need to share memory layout with the build script, LUA scripts and the assemblers (DASM & ACME).
We could define these addresses in the build script and pass them into the programs, but this would take a lot of work and I won't persue this now.
(The LUA scripts generate a lot of assembler files to wrap binaries into PRG files -- this will go away with the fast-loader able to handle raw data)