Fastload the SID on the logo screen
It's not played yet; at the moment the SID is bundled into the outfit binary but our fast-loader doesn't support multi-part exo files so we will have to load multiple files in sequence.

I would like to havd the music playing as early as possible during the splash and loading of the main outfit.